+++ InheritanceTask/InheritanceTask/DocumentProgram.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace InheritanceTask
+{
+    class DocumentProgram
+    {
+        public void OpenDocument()
+        {
+            Console.WriteLine("Document Opened");
+        }
+
+        public virtual void EditDocument()
+        {
+            Console.WriteLine("Can Edit in Pro and Expert versions");
+        }
+
+        public virtual void SaveDocument()
+        {
+            Console.WriteLine("Can Save in Pro and Expert versions");
+        }
+    }
+}
