+++ InheritanceTask/InheritanceTask/ExpertDocument.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace InheritanceTask
+{
+    class ExpertDocument : ProDocumentProgram
+    {
+        public override void SaveDocument()
+        {
+            Console.WriteLine("Document Saved in pdf format");
+        }
+    }
+}
