+++ InheritanceTask/InheritanceTask/ProDocumentProgram.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace InheritanceTask
+{
+    class ProDocumentProgram : DocumentProgram
+    {
+        public override sealed void EditDocument()
+        {
+            Console.WriteLine("Document Edited");
+        }
+
+        public override void SaveDocument()
+        {
+            Console.WriteLine("Document Saved in doc format, for pdf format buy Expert packet");
+        }
+    }
+}
